Moved conversation update subscription so it is attached on resumed conversations not just new ones
